+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `init` command now shows progress while it creates files and runs final setup commands
+
 ## [0.3.0] - 2026-04-06
 
 ### Added